auth_challenge_server_python/requirements.txt
fastmcp>=0.1.0
mcp>=0.1.0
python-dotenv>=1.0.1
uvicorn>=0.30.0
